refactor: move dock host API to workflow module
